refactor(api): modularize location controllers and include iso code
