+++ api/src/controllers/location/setCountries.js
@@ -1,6 +1,6 @@
 const fs = require('fs');
 const path = require('path');
-const { Countries } = require('../db')
+const { Countries } = require('../../db')
 
 async function readCountries(filePath) {
     try {
@@ -11,6 +11,9 @@ async function readCountries(filePath) {
             await Countries.findOrCreate({
                 where: {
                     name: country.name
+                }, 
+                defaults: {
+                    iso: country.iso2
                 }
             });
         }
@@ -21,7 +24,7 @@ async function readCountries(filePath) {
 
 async function setCountries(req, res) {
     try {
-        const filePath = path.join(__dirname, '../utils/countries.json')
+        const filePath = path.join(__dirname, '../../utils/countries.json')
         await readCountries(filePath);
         res.status(200).json("Countries saved successfully")
     } catch (error) {
