+++ api/src/models/Countries.js
@@ -10,6 +10,10 @@ module.exports = (sequelize) => {
         name: {
             type: DataTypes.STRING,
             allowNull: false
+        },
+        iso: {
+            type: DataTypes.STRING,
+            allowNull: false
         }
     }, {
         timestamps: false
